move payloadvariables to parser
