+++ src/engine/SpatialJoin.h
@@ -10,8 +10,8 @@
 #include <variant>
 
 #include "engine/Operation.h"
-#include "engine/PayloadVariables.h"
 #include "global/Id.h"
+#include "parser/PayloadVariables.h"
 #include "parser/data/Variable.h"
 
 // A nearest neighbor search with optionally a maximum distance.
